refactor(ui): icon namespace
