+++ packages/photo-companion/src/routes/height-by-shadow/height-by-shadow.page.tsx
@@ -1,4 +1,4 @@
-import { IconShadow } from "ui";
+import { IconShadow } from "ui/icons";
 
 import { useTranslation } from "~/services/translation";
