+++ packages/photo-companion/src/routes/settings/components/input-tabs-select/input-tabs-select.tab.tsx
@@ -1,4 +1,5 @@
-import { Button, IconArrowDown, IconArrowUp, InputCheckbox } from "ui";
+import { Button, InputCheckbox } from "ui";
+import { IconArrowDown, IconArrowUp } from "ui/icons";
 
 import { type ROUTE_PRIMARY_LABEL } from "~/services/navigation";
 import { useTranslation } from "~/services/translation";
