+++ packages/ui/src/components/index.ts
@@ -5,29 +5,6 @@ export { Disclose } from "./disclose/disclose";
 export { Drawer } from "./drawer/drawer";
 export { Fieldset } from "./form/fieldset";
 export { Form } from "./form/form";
-export {
-	IconAnchor,
-	IconArrowDown,
-	IconArrowUp,
-	IconBook,
-	IconBridge,
-	IconChevronDown,
-	IconChevronLeft,
-	IconChevronRight,
-	IconClose,
-	IconCube,
-	IconGithub,
-	IconLights,
-	IconMenu,
-	IconMoon,
-	IconPullRequest,
-	IconQuestion,
-	IconSettings,
-	IconShadow,
-	IconSun,
-	IconTimeline,
-	IconWarning
-} from "./icon";
 export { Icon, type IconProps } from "./icon/icon";
 export { InputCheckbox } from "./input-checkbox/input-checkbox";
 export { InputDatetime } from "./input-datetime/input-datetime";
