--- packages/ui/src/index.ts
@@ -1,16 +1,0 @@
-export {
-	Button,
-	Details,
-	Fieldset,
-	Form,
-	Icon,
-	Input,
-	InputCheckbox,
-	InputDatetime,
-	InputRadio,
-	InputSelect,
-	type InputSelectOption,
-	Link,
-	Loader,
-	Time
-} from "./components";
